+++ kinto/views/groups.py
@@ -5,9 +5,15 @@ from kinto.core.events import ResourceChanged, ACTIONS
 from pyramid.events import subscriber
 
 
+def validate_member(node, member):
+    if member.startswith('/buckets/') or member == 'system.Everyone':
+        raise colander.Invalid(node, "%r is not a valid user ID." % member)
+
+
 class GroupSchema(resource.ResourceSchema):
     members = colander.SchemaNode(colander.Sequence(),
-                                  colander.SchemaNode(colander.String()))
+                                  colander.SchemaNode(colander.String(),
+                                                      validator=validate_member))
 
 
 @resource.register(name='group',